+++ templates/templates-client/src/main/java/io/lanu/warmsnow/templates/templates_client/dto/FieldDto.java
@@ -15,7 +15,7 @@ public class FieldDto {
     private int position;
     private int level;
     private FieldType fieldType;
-    private BigDecimal productivity;
+    private int productivity;
     private boolean underUpgrade;
     private boolean ableToUpgrade;
     private long timeToNextLevel;
