+++ templates/templates-server/src/main/java/io/lanu/warmsnow/templates/templates_server/entities/FieldTemplateEntity.java
@@ -17,7 +17,7 @@ public class FieldTemplateEntity {
     private int position;
     private int level;
     private FieldType fieldType;
-    private BigDecimal productivity;
+    private int productivity;
     private boolean underUpgrade;
     private boolean ableToUpgrade;
     private long timeToNextLevel;
